requirements.txt
streamlit
pandas
Aggrid